Refactoring to create valid project/scope to include a target
Tutorials need to be updated in order to work. Will need to go find where the boilerplate config gets generated too.

I suspect I'll find that I'll be able to do a run overriding the target when one is not defined in the scope or pipeline, but so far the basic tutorial doesn't get into that.